// src/types/ISearchProducts.ts
export interface ISearchProducts {
    films: [],
    kinopoiskId: number;
    nameRu: string;
    posterUrl: string;
    posterUrlPreview: string,
    ratingKinopoisk: number;
    year: number;
    filmLength: number;
    slogan: string;
    description: string;
    shortDescription: string;
    countries: string;
    keyword: string
  }
  